Adds new supported game platforms
Adds support for Oculus Quest, Oculus VR, Steam VR, PSVR, PSVR2, and Vita platforms to the `SupportedGamePlatform` enum.

Also, it ensures `PlatformChip` is only displayed if platform data is available and uses 'TBD' as a default label for game dates when the human-readable date is null.

Additionally, it improves scroll physics by safely casting to RenderBox.